requirements.txt
requests==2.31.0
pytest==8.1.1
furl==2.1.3
torch==2.2.1
Flask==3.0.2
pillow==10.2.0
torchvision==0.17.1
grpcio==1.62.1
grpcio-tools==1.62.1
prometheus-flask-exporter==0.23.0
prometheus_client==0.20.0
supervisor==4.2.5